+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The present file will list all changes made to the project; according to the
 
 ### Added
 - Sessions tab for OAuth Clients to display non-expired sessions associated with the client and allow revoking them.
+- Improved client IP detection.
+  If your GLPI instance is behind a reverse proxy, you should add its IP(s) the new `GLPI_TRUSTED_REVERSE_PROXIES` constant and modify the new `GLPI_REVERSE_PROXY_HEADERS` constant to include the headers your proxy uses to forward the client IP.
+  Only the required HTTP headers should be listed for better security as any header not handled by the proxy could be spoofed by the client.
 
 ### Changed
 - "Computer" search option (ID 12) for Databases has been replaced by "Associated item type" (ID 14) and "Associated item" (ID 12) options. These are not searchable but can be displayed.